fix: prevent integer overflow in chat template functions
Change return type and length parameter from int32_t to int64_t in
chat template functions to prevent overflow when processing large
inputs (>2GB).

When buffer size exceeded INT32_MAX, dest.size() would overflow to a
negative value when cast to int32_t, causing a misleading "template
not supported" error instead of properly handling large inputs.

Files changed:
- include/llama.h: Update API signature
- src/llama.cpp: Update implementation
- src/llama-chat.h: Update internal declaration
- src/llama-chat.cpp: Update internal implementation
- common/chat.cpp: Update result variable type

Fixes #17463

+++ common/chat.cpp
@@ -2829,7 +2829,7 @@ static common_chat_params common_chat_templates_apply_legacy(
 
     // run the first time to get the total output length
     const auto & src = tmpls->template_default->source();
-    int32_t res = llama_chat_apply_template(src.c_str(), chat.data(), chat.size(), inputs.add_generation_prompt, buf.data(), buf.size());
+    int64_t res = llama_chat_apply_template(src.c_str(), chat.data(), chat.size(), inputs.add_generation_prompt, buf.data(), buf.size());
 
     // error: chat template is not supported
     if (res < 0) {
